Refactor to allow for V1 or V2 of GPACT protocol
